[RF][HS3] Move HS3 HistFactory implementation to `histfactory`
This will make it easier to merge the old HistFactory implementation and
the new HistFactory implementation from RooFitHS3.
