+++ roofit/hs3/LinkDef.h
@@ -3,7 +3,6 @@
 #pragma link off all classes;
 #pragma link off all functions;
 
-#pragma link C++ class RooStats::HistFactory::JSONTool + ;
 #pragma link C++ class RooJSONFactoryWSTool + ;
 #pragma link C++ class RooFit::JSONIO::Importer + ;
 #pragma link C++ class RooFit::JSONIO::Exporter + ;
