+++ roofit/hs3/src/JSONIOUtils.cxx
@@ -3,16 +3,6 @@
 using RooFit::Detail::JSONNode;
 using RooFit::Detail::JSONTree;
 
-bool startsWith(std::string_view str, std::string_view prefix)
-{
-   return str.size() >= prefix.size() && 0 == str.compare(0, prefix.size(), prefix);
-}
-
-bool endsWith(std::string_view str, std::string_view suffix)
-{
-   return str.size() >= suffix.size() && 0 == str.compare(str.size() - suffix.size(), suffix.size(), suffix);
-}
-
 std::unique_ptr<RooFit::Detail::JSONTree> varJSONString(const JSONNode &treeRoot)
 {
    std::string varName = treeRoot.find("name")->val();
