+++ roofit/roofitcore/src/RooFitImplHelpers.cxx
@@ -30,6 +30,16 @@
 
 #include <unordered_map>
 
+bool startsWith(std::string_view str, std::string_view prefix)
+{
+   return str.size() >= prefix.size() && 0 == str.compare(0, prefix.size(), prefix);
+}
+
+bool endsWith(std::string_view str, std::string_view suffix)
+{
+   return str.size() >= suffix.size() && 0 == str.compare(str.size() - suffix.size(), suffix.size(), suffix);
+}
+
 namespace RooHelpers {
 
 namespace {
